Implemented the .goalRegion functionality. An object with this name suffix will be automatically converted into a collision sensor and the goal will not be satisfied unless the corresponding rigid body is colliding with it. This allows the user to specify an arbitrarily shaped region that an object must touch in order to satisfy the goal.
Moved some of the goal setup code from builder.py to the Plan operator in ompl_addon.py.in. The changes performed automatically are thus saved to the .blend file so they don't have to be re-performed every time unless the user destroys them.

Included a new bgetest2.blend file illustrating the use of .goalRegion objects.