---------------------------------------------------- Changes in FMT.h: ---------------------------------------------------- Swapped the parameter 'obstacleCoverage_' for 'freeSpaceVolume_'. Ideally, we would like to use a variable called 'freeSpaceFraction_,' that denotes the fraction of the configuration space that is free of obstacles. If the state space class had a 'getVolume()' method, then we could implement this alternative FMT class variable and make the algorithm interface easier for the user.
Added a 'maxSampleAttempts_' class variable.

Added a 'lastGoalMotion_' class variable.

Changed the radiusMultiplier_ lower bound to 0 instead of 1.

Removed the class variable 'children' from the Motion class.

Minor stylistic changes.

----------------------------------------------------
Changes in FMT.cpp:
----------------------------------------------------
In the constructor -
Swapped the parameter 'obstacleCoverage_' for 'freeSpaceVolume_.'
Added the 'maxSampleAttempts_' variable.
Added the 'lastGoalMotion_' variable.
Updated the planner parameter suggested ranges.

In solve -
Set the variable lastGoalMotion_ at the end of solve().

In getPlannerData-
Handled start and goal states differently than other states.

In assureGoalIsSampled:
Utilized the nextGoal() function in a while loop, in place of the for loop that was previously implemented.

Minor stylistic changes.