Add option to skip setting RPATH
OMPL installs to /usr/lib by default and some projects integrating
OMPL trigger a QA check suggesting that the set RPATH is useless.
For example Yocto gives this warning:

    QA Issue: ompl: /work/corei7-64-refkit-linux/ompl/1.1.0-r0/packages-split/ompl/usr/lib/libompl.so.1.1.0 contains probably-redundant RPATH /usr/lib [useless-rpaths]

Add the option OMPL_SKIP_RPATH that helps to skip setting RPATH
for the library. By default the option is OFF.