Implement repairing search upon decrease of inflation factor
To achieve proper ARA* like behaviour, we need to keep track of the
set of inconsistent vertices. That's because upon researching the same
approximation with a new inflation factor, we don't have to begin from
the start vertex. We only have to consider the edges that are still
open, and the ones that are outgoing of inconsistent vertices.