Let edges be between states, not vertices
This not only gets rid of a lot of pointer directions but also cleans up the code conceptually. Now,
every vertex except the roots must have a parent (and only a parent). Actually this makes it sound
like the children should be std::unique_ptr's. Let me think about that a little more.