Update the pruning criterion
Profiling reveals that pruning is very expensive, mainly because the
nearest neighbour structure of the samples has to be updated. On the
other hand, nearest neighbour lookup gets more expensive the bigger
the structure, so it's a tradeoff. Pruning on every cost update seems
insensible, but so does never pruning at all. The criteria to prune
should depend on how many vertices/samples there are and how many of
them could be pruned, as the decrease in cost associated with nearest
neighbour lookup for fewer samples must justify the cost of pruning.
It turns out that counting is affordable, so we don't need to use any
proxy here.