Improve the sparse collision detection algorithm in EIT*
Let's say I have an edge like this

         |--------------------------------------------------------------------|

Before this commit, the collision checking resolution was increased like this:

Level 0: |--------------------------------------------------------------------|
Level 1: |---------------------------------x----------------------------------|
Level 2: |----------------------x---------------------x-----------------------|
Level 3: |---------------x-----------------x-----------------x----------------|
Level 4: |------------x-------------x-------------x-------------x-------------|
Level 5: |---------x-----------x-----------x-----------x------------x---------|

But this means that by the time I get to, say, level 5, the distribution of checks looks like this:

Level 5: |---------x--x--x-----xx---x-----3x------x---xx-----x--x---x---------|

This is not exactly uniform. Instead the collision checking resolution is now like this:

Level 0: |--------------------------------------------------------------------|
Level 1: |---------------------------------x----------------------------------|
Level 2: |---------------x-----------------------------------x----------------|
Level 3: |-------x----------------x-----------------x-----------------x-------|

This means that already at level 3 we have the following distribution of checks:

Level 3: |-------x-------x--------x--------x--------x--------x--------x-------|

Compared to level 3 before, with only 1 check more.

Level 3: |---------------x------x---------2x----------x------x----------------|

One drawback of this approach is that some edges might not have been checked at previous levels, and
therefore the will not contain all checks. But since every level has a well distributed set of
checks on its own, I don't think this is a big deal. But it means that we can never whitelist an
edge if it passes a certain level, because we don't keep track of whether the edge also passed all
levels below.