Ensure registration of adding children to vertices is done correctly
Registering the expansion was a bit of a misnomer. Expanding means getting a vertice's neighbors,
and that's actually not what's relevant here. We need to make sure that a vertices cost-to-come
doesn't change once it has a child added to it, as this would invalidate the cost-to-come of the children.