Don't reset the k-nearest RGG constant when clearing the graph
Clear() should reset the planner to the state it had before it started solving the problem, not the
state it had before receiving the problem definition. Since the k-nearest RGG constant only depends
on the state space dimension (which is part of the problem definition), it should not get reset when
clear is called.