Fix/aitstar miscellaneous (#763)
* Fix neighbor cache in AIT*

Switch to storing cached neighbors in weak pointers, as shared pointers can result in memory leaks.

* Fix AIT*'s caching of whitelisted children

The cache sometimes held weak pointers to vertices that were once whitelisted as children but have
since been pruned. This commit fixes that by checking whether the vertex is still valid and removing
it from the cache if not.

* Add AIT* to optimal planning demo

Add 'AITstar' as an option for the optimal planning demo.

* Check termination condition between each successful sample

If sampling valid states is expensive, AIT* does not respect the timeout conditions otherwise.