Decrease initial suboptimality factor and remove option to reset it
Infinity is a bad idea for the initial suboptimality factor as that will make
all edges in the queue look equally promising. The maximum double is also a bad
idea, as the suboptimality factor is multiplied with the cost-to-go and if that
cost is greater than 1, then it will result in infinity and we have the same
problem as before.