Fix collision detection in reverse search
This fixes a bug in the sparse collision detection of the reverse search of
EIT*. Prior to this commit, the code implicitly assumed that all edges in the
graph had already been checked at the resolution of the last reverse search,
even though only the edges that were processed by the reverse search have
actually been checked. This commit adds a lookup to each state that caches the
collision detection resolution at which incoming edges were checked and skips
those checks when either checking the edge sparsely with higher resolution in
the reverse search and when checking the edge at full resolution in the forward
search.