Set the sampled cost to identity when adding a new batch in BIT*
Here's how I understand what the problem was: If sampling from a compound state
space with RealVectorStateSpaces as components, then direct informed sampling
is used for these components. To support just-in-time sampling, BIT* samples
shells of the informed set in these components. These shells are defined by a
minimum and maximum cost. The issue with initializing sampledCost_ as the
minimum possible cost to solve the problem is that some of that cost might come
from non-RealVectorStateSpace components of the compound state space, e.g., the
angular component in SE(2). If this is the case, then the center of the
informed set in the RealVectorStateSpace components is never sampled.

We could initialize sampledCost_ to the smallest minimum possible cost of each
RealVectorStateSpace component of the compound space, but that is nontrivial to
figure out and also not necessary. Instead we can just initialize sampledCost_
to identity, which ensures that the center of all informed sets in all
RealVectorStateSpaces is sampled.