Extend key of reverse queue with cost to target
If two edges in the reverse queue have equal total potential solution cost,
then we should tiebreak first on the cost to the target of the edge to save a
rewiring later. This almost never happens though because it's highly unlikely
that two edges have equal total potential solution cost unless the cost
heuristic is zero in which case they will also have equal cost to the target of
the edge and we're tiebreaking on the effort anyway.