Consider reverse search vertices without potential children closed
Prior to this change: If a vertex cannot have any children, because it is not
close enough to any state except its parent, then it will never be considered
closed because vertices are considered closed when their first outgoing edge is
processed.

With this change: Vertices without any potential children are considered closed
when they are encountered. From an ordering perspective, this is equivalent to
the vertex having a potential child that is infinitely close to it, which EIT*
can handle.

I think this is conceptually correct and it satisfies the assert that checks
that all processed edges in the forward queue have target states that are
closed in the reverse search.