Use more conservative types in PLY files
In PLY files it is more common and well supported to use a uchar for the
number of face vertices and an int for the vertex indices. These are
also the types used in the PLY format description. Certain software
(e.g. Meshlab) doesn't support using a uint for either type.

http://paulbourke.net/dataformats/ply/