ensure that `webbrowser` pkg will not use terminal browser
for register step. If DISPLAY is not set, usually the command
would invoke a browser like `lynx`, `elinks`, etc.
However, that does not work because (1) the command output is quiet,
(2) cmake controls the shell, and (3) the browser does not quit without
user input anymore blocking the build. With the overridden TERM variable
the process will simply fail and the build goes on.